Introduce SwaggerTag combinator
