+++ libs/wire-api/src/Wire/API/Routes/Internal/Brig.hs
@@ -60,6 +60,7 @@ import Wire.API.Routes.Internal.Brig.EJPD
 import qualified Wire.API.Routes.Internal.Galley.TeamFeatureNoConfigMulti as Multi
 import Wire.API.Routes.MultiVerb
 import Wire.API.Routes.Named
+import Wire.API.SwaggerServant
 import Wire.API.Team.Feature
 import Wire.API.User
 import Wire.API.User.Auth
@@ -319,7 +320,8 @@ type GetVerificationCode =
     :> Get '[Servant.JSON] (Maybe Code.Value)
 
 type API =
-  "i"
+  SwaggerTag "brig"
+    :> "i"
     :> ( EJPD_API
            :<|> AccountAPI
            :<|> MLSAPI
