+++ libs/wire-api/wire-api.cabal
@@ -116,6 +116,7 @@ library
     Wire.API.Routes.WebSocket
     Wire.API.ServantProto
     Wire.API.SwaggerHelper
+    Wire.API.SwaggerServant
     Wire.API.SystemSettings
     Wire.API.Team
     Wire.API.Team.Conversation
